+++ backend/app/controllers/comable/admin/products_controller.rb
@@ -7,14 +7,16 @@ module Comable
 
       def index
         @q = Comable::Product.ransack(params[:q])
-        @products = @q.result(distinct: true).includes(:stocks, :images).page(params[:page]).accessible_by(current_ability)
+        @products = @q.result(distinct: true).includes(:images, variants: [:option_values, :stock]).page(params[:page]).accessible_by(current_ability)
       end
 
       def show
+        edit
         render :edit
       end
 
       def new
+        @product.variants.build
         @product.published_at = Date.today
       end
 
@@ -28,6 +30,7 @@ module Comable
       end
 
       def edit
+        set_preview_session
       end
 
       def update
@@ -66,21 +69,23 @@ module Comable
 
       private
 
-      # rubocop:disable Metrics/MethodLength
       def product_params
         params.require(:product).permit(
           :name,
-          :code,
           :caption,
-          :price,
+          :property,
           :published_at,
-          :sku_h_item_name,
-          :sku_v_item_name,
           category_path_names: [],
-          images_attributes: [:id, :file, :_destroy]
+          images_attributes: [:id, :file, :_destroy],
+          variants_attributes: [:id, :price, :sku, :options, :quantity, :_destroy],
+          option_types_attributes: [:id, :name, { values: [] }]
         )
       end
-      # rubocop:enable Metrics/MethodLength
+
+      def set_preview_session
+        session[Comable::Product::PREVIEW_SESSION_KEY] ||= {}
+        session[Comable::Product::PREVIEW_SESSION_KEY][@product.to_param] = true
+      end
     end
   end
 end