+++ backend/config/routes.rb
@@ -5,6 +5,8 @@ Comable::Core::Engine.routes.draw do
     resource :dashboard, only: :show
 
     resources :products do
+      resources :variants
+
       resources :stocks
 
       collection do
@@ -59,7 +61,7 @@ Comable::Core::Engine.routes.draw do
 
     resource :store, controller: :store, only: [:show, :edit, :update]
 
-    devise_for :user, path: :user, class_name: Comable::User.name, module: :devise, controllers: {
+    devise_for :users, path: :user, class_name: Comable::User.name, module: :devise, router_name: :comable, controllers: {
       sessions: 'comable/admin/user_sessions'
     }