+++ core/app/controllers/comable/application_controller.rb
@@ -1,0 +1,4 @@
+module Comable
+  class ApplicationController < ActionController::Base
+  end
+end