+++ core/app/models/comable/product.rb
@@ -6,21 +6,22 @@ module Comable
     include Comable::Product::Search
     include Comable::Product::Csvable
 
-    has_many :stocks, class_name: Comable::Stock.name, dependent: :destroy
+    has_many :variants, class_name: Comable::Variant.name, inverse_of: :product, dependent: :destroy
     has_many :images, class_name: Comable::Image.name, dependent: :destroy
     has_and_belongs_to_many :categories, class_name: Comable::Category.name, join_table: :comable_products_categories
 
+    accepts_nested_attributes_for :variants, allow_destroy: true
     accepts_nested_attributes_for :images, allow_destroy: true
 
+    scope :published, -> (published_at = nil) { where('published_at <= ?', published_at || Time.now) }
+
     validates :name, presence: true, length: { maximum: 255 }
-    validates :code, presence: true, length: { maximum: 255 }
-    validates :price, presence: true, numericality: { greater_than_or_equal_to: 0, allow_blank: true }
-    validates :sku_h_item_name, length: { maximum: 255 }
-    validates :sku_v_item_name, length: { maximum: 255 }
 
     liquid_methods :id, :code, :name, :price, :images, :image_url
 
-    ransack_options attribute_select: { associations: :stocks }
+    ransack_options attribute_select: { associations: [:variants, :stocks, :option_types] }
+
+    PREVIEW_SESSION_KEY = :preview_product
 
     # Add conditions for the images association.
     # Override method of the images association to support Rails 3.x.
@@ -50,10 +51,67 @@ module Comable
       self.categories = Comable::Category.find_by_path_names(category_path_names, delimiter: delimiter)
     end
 
-    private
-
-    def create_stock
-      stocks.create(code: code) unless stocks.exists?
+    def master?
+      option_types.empty?
     end
+
+    def as_json(options = nil)
+      super (options || {}).merge(methods: [:variants])
+    end
+
+    def sku_h_item_name
+      option_types.first.try(:name)
+    end
+
+    def sku_v_item_name
+      option_types.second.try(:name)
+    end
+
+    def code
+      variants.first.try(:sku)
+    end
+
+    def code=(code)
+      variants.each { |v| v.sku = code }
+    end
+
+    def price
+      variants.first.try(:price)
+    end
+
+    def price=(price)
+      variants.each { |v| v.price = price }
+    end
+
+    has_many :stocks, class_name: Comable::Stock.name, through: :variants
+
+    def stocks=(stocks)
+      stocks.map { |stock| variants.build(stock: stock) }
+    end
+
+    def option_types
+      variants.map { |variant| variant.option_values.map(&:option_type) }.flatten.uniq
+    end
+
+    def option_types_attributes=(_option_types_attributes)
+    end
+
+    def properties
+      parse_property = JSON.parse(property) # propertyに不正な値が入っていた場合に例外発生する
+      return [] unless parse_property.is_a?(Array) && parse_property.all? { |prop| prop.is_a?(Hash) }
+      parse_property
+    rescue
+      []
+    end
+
+    #
+    # Deprecated methods
+    #
+    deprecate :stocks, deprecator: Comable::Deprecator.instance
+    deprecate :sku_h_item_name, deprecator: Comable::Deprecator.instance
+    deprecate :sku_v_item_name, deprecator: Comable::Deprecator.instance
+    deprecate :code, deprecator: Comable::Deprecator.instance
+    deprecate :code=, deprecator: Comable::Deprecator.instance
+    deprecate :option_types_attributes=, deprecator: Comable::Deprecator.instance
   end
 end