+++ core/app/models/comable/stock/csvable.rb
@@ -7,19 +7,18 @@ module Comable
 
       included do
         comma do
-          product_code
-          code
+          product_id
+          id
           quantity
           sku_h_choice_name
           sku_v_choice_name
         end
       end
 
-      delegate :code, to: :product, prefix: true, allow_nil: true
+      delegate :id, to: :product, prefix: true, allow_nil: true
 
-      def product_code=(code)
-        return if product_code == code
-        self.product = Comable::Product.find_by(code: code)
+      def product_id=(id)
+        self.product = Comable::Product.find_by(id: id)
       end
     end
   end