+++ core/db/migrate/20150918065115_add_property_to_comable_products.rb
@@ -1,0 +1,5 @@
+class AddPropertyToComableProducts < ActiveRecord::Migration
+  def change
+    add_column :comable_products, :property, :text, after: :caption
+  end
+end