+++ spec/factories/comable/option_values.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :option_value, class: Comable::OptionValue do
+    sequence(:name) { |n| "Option Value ##{n.next}" }
+  end
+end