+++ spec/rails_helper.rb
@@ -12,6 +12,15 @@ require 'rspec/example_steps'
 require 'shoulda/matchers'
 require 'generator_spec'
 
+# for Feature test
+require 'capybara'
+require 'phantomjs'
+require 'phantomjs/poltergeist'
+require 'database_cleaner'
+
+# Change Capybara javascript driver to Poltergeist (PhantomJS)
+Capybara.javascript_driver = :poltergeist
+
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.
 Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
@@ -24,15 +33,12 @@ RSpec.configure do |config|
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"
 
-  # If you're not using ActiveRecord, or you'd prefer not to run each of your
-  # examples within a transaction, remove the following line or assign false
-  # instead of true.
-  config.use_transactional_fixtures = true
-
+  # Modules
   config.include Comable::EngineControllerTestMonkeyPatch, type: :controller
   config.include Devise::TestHelpers, type: :view
   config.include Devise::TestHelpers, type: :controller
-  config.extend Comable::AuthHelpers, type: :controller
+  config.extend Comable::AuthorizationHelpers::Controller, type: :controller
+  config.extend Comable::AuthorizationHelpers::Feature, type: :feature
   config.extend Comable::RequestHelpers, type: :request
 
   # for Rspec 3
@@ -41,4 +47,28 @@ RSpec.configure do |config|
 
   # Omit the prefix FactoryGirl
   config.include FactoryGirl::Syntax::Methods
+
+  # Support Capybara DSL for the feature test
+  config.include Capybara::DSL, type: :feature
+
+  #
+  # DatabaeeCleaner for the asynchronous test (with :js option).
+  #
+  config.use_transactional_fixtures = false
+
+  # Clean up database before test
+  config.before(:suite) do
+    DatabaseCleaner.clean_with(:truncation)
+  end
+
+  # Start the transaction
+  config.before(:each) do |example|
+    DatabaseCleaner.strategy = example.metadata[:js] ? :truncation : :transaction
+    DatabaseCleaner.start
+  end
+
+  # Clean up database by the strategy
+  config.after(:each) do
+    DatabaseCleaner.clean
+  end
 end